Fix type of costs

+++ backend/api/migrations/0003_manual_view.py
@@ -257,7 +257,7 @@ class Migration(migrations.Migration):
                 cell_saver_ml MEDIUMINT UNSIGNED DEFAULT 0,
                 overall_units SMALLINT UNSIGNED AS (rbc_units + ffp_units + plt_units + cryo_units + whole_units) STORED,
 
-                los DECIMAL(6,2),
+                los FLOAT,
                 death BOOLEAN,
                 vent BOOLEAN,
                 stroke BOOLEAN,
